requirements.txt
beautifulsoup4==4.11.1
mysql_connector_repackaged==0.3.1
pandas~=1.5.2
requests==2.28.1
tqdm==4.64.1

mysql-connector-python~=8.0.31
